test-framework/requirements.txt
pytest==7.2.1
pytest-html==3.2.0
scapy==2.4.5
confluent_kafka==2.0.2
requests==2.31.0
avro==1.11.3
fastavro==1.9.3
prettytable==3.4.0
PyYAML==6.0.1
wcwidth==0.2.5
 